requirements.txt
bcrypt
flask
flask_socketio
peewee
requests